+++ private/js/tiptap_plugins/cms.tiptap.toolbar.js
@@ -4,36 +4,46 @@
 
 'use strict';
 
+import { getDefaultTableClass } from "./cms.table";
+
 
 function generateButtonArray(rows, cols) {
-    const buttons = ['<div class="tt-create-table">'];
+    const buttons = [];
     for (let j= 0; j < rows; j++) {
         for (let i = 0; i < cols; i++) {
             buttons.push(`<button title="${i+1}x${j+1}" data-action="Table" style="--mx: ${i*12}px; --my: ${j*12+4}px;" data-rows="${j+1}" data-cols="${i+1}"></button>`);
         }
     }
-    buttons.push('</div>');
-    return buttons.join('');
+    return `<div class="tt-create-table">${buttons.join('')}</div>`;
 }
 
 const _tableMenu = [
-    'addColumnBefore',
-    'addColumnAfter',
-    'deleteColumn',
-    '|',
-    'addRowBefore',
-    'addRowAfter',
-    'deleteRow',
-    '|',
-    'toggleHeaderColumn',
-    'toggleHeaderRow',
-    'mergeOrSplit',
+    ['addColumnBefore', 'addColumnAfter', 'deleteColumn'],
+    ['addRowBefore', 'addRowAfter', 'deleteRow'],
+    ['toggleHeaderColumn', 'toggleHeaderRow', 'mergeOrSplit']
 ];
 
-function generateTableMenu(editor, builder) {
-    return generateButtonArray(10, 10) + builder(_tableMenu);
+function generateTableMenu(editor, builder, item) {
+    let tableMenu = builder(_tableMenu);
+
+    if (tableMenu.endsWith(editor.options.separator_markup)) {
+        tableMenu = tableMenu.slice(0, -editor.options.separator_markup.length);
+    }
+    return generateButtonArray(10, 10)
+        + `<div class="tt-edit-table">${tableMenu}</div>`
+        + builder(['tableClass']);
 }
 
+function renderTableClassOptions(editor, item) {
+    const tableClasses = editor.options.tableClasses || cms_editor_plugin.tableClasses;
+    if (!Array.isArray(tableClasses)) {
+        // Only one option - no need to render a button to select it
+        return '';
+    }
+    return '<div class="hr"></div>' + tableClasses
+        .map((cls) => `<button data-action="tableClass" data-attr="${cls[0]}">${cls[1] || cls[0]}</button>`)
+        .join('');
+}
 
 const TiptapToolbar = {
     Undo: {
@@ -192,18 +202,6 @@ const TiptapToolbar = {
         active: (editor) => editor.isActive('bulletList'),
         type: 'block',
     },
-/*
-    Outdent: {
-        action: (editor) => editor.chain().focus().outdent().run(),
-        enabled: (editor) => editor.can().outdent(),
-        type: 'block',
-    },
-    Indent: {
-        action: (editor) => editor.chain().focus().indent().run(),
-        enabled: (editor) => editor.can().indent(),
-        type: 'block',
-    },
-*/
     Blockquote: {
         action: (editor) => editor.chain().focus().toggleBlockquote().run(),
         enabled: (editor) => editor.can().toggleBlockquote(),
@@ -247,7 +245,8 @@ const TiptapToolbar = {
         action: (editor, button) => {
             const rows = parseInt(button?.dataset?.rows ||3);
             const cols = parseInt(button?.dataset?.cols || 3);
-            editor.chain().focus().insertTable({ rows: rows, cols: cols }).run();
+            const classes = button?.dataset?.attr || getDefaultTableClass(editor.options.tableClasses);
+            editor.chain().focus().insertTable({ rows: rows, cols: cols}).updateAttributes('table', { addClasses: classes }).run();
         },
         enabled: (editor) => editor.can().insertTable({ rows: 3, cols: 3 }),
         type: 'mark',
@@ -300,6 +299,18 @@ const TiptapToolbar = {
         action: (editor, button) => editor.chain().focus().mergeOrSplit().run(),
         enabled: (editor) => editor.can().mergeOrSplit(),
         type: 'mark',
+    },
+    tableClass: {
+        action: (editor, button) => {
+            const classes = button?.dataset.attr || getDefaultTableClass(editor.options.tableClasses);
+            editor.chain().focus().updateAttributes('table', { addClasses: classes }).run();
+        },
+        enabled: (editor, button) => editor.can().updateAttributes('table', { addClasses: button?.dataset.attr || ''}),
+        active: (editor, button) => {
+            const classes = editor.getAttributes('table').addClasses;
+            return classes && classes === button.dataset.attr;
+        },
+        render: renderTableClassOptions,
     },
     Code: {
         action: (editor) => editor.chain().focus().toggleCode().run(),